--- aula-1/POO_exemplos/Com_heranca/carro.js
@@ -1,12 +1,0 @@
-import { Veiculo } from "./veiculo";
-
-export class Carro extends Veiculo {
-    
-    constructor(nome, marca, ano, preco) {
-        super(nome, marca, ano, preco);
-    }
-
-    abrirPorta() {
-        console.log("Porta aberta!");
-    }
-}
