--- aula-1/POO_exemplos/Com_heranca/main.js
@@ -1,5 +1,0 @@
-import { Carro } from "./carro";
-import { Moto } from "./moto";
-
-carro = new Carro();
-moto = new Moto();
