--- aula-1/POO_exemplos/Com_heranca/moto.js
@@ -1,12 +1,0 @@
-import { Veiculo } from "./veiculo";
-
-export class Moto extends Veiculo {
-    
-    constructor(nome, marca, ano, preco) {
-        super(nome, marca, ano, preco);
-    }
-    
-    empinar() {
-        console.log("Mandando grau!");
-    }
-}
