--- aula-1/POO_exemplos/Sem_heranca/carro.js
@@ -1,25 +1,0 @@
-export class Carro {
-    nome;
-    marca;
-    ano;
-    preco;
-
-    constructor(nome, marca, ano, preco) {
-        this.nome = nome;
-        this.marca = marca;
-        this.ano = ano;
-        this.preco = preco;
-    }
-
-    acelerar() {
-        console.log("Acelerando!");
-    }
-
-    parar() {
-        console.log("Freios acionados!");
-    }
-
-    abrirPorta() {
-        console.log("Porta aberta!");
-    }
-}
