--- aula-1/POO_exemplos/Sem_heranca/main.js
@@ -1,12 +1,0 @@
-import { Carro } from "./carro";
-import { Moto } from "./moto";
-
-carro = new Carro();
-moto = new Moto();
-
-
-carro.acelerar();
-carro.parar();
-carro.abrirPorta();
-
-moto.empinar();
